chore: refactor architecture
