+++ src/main/java/com/supplyhub/domain/employee/EmployeeRepository.java
@@ -1,0 +1,8 @@
+package com.supplyhub.domain.employee;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+    Employee findByCpf(String cpf);
+}
