+++ src/main/java/com/supplyhub/domain/employee/dto/CreateDataEmployeeDto.java
@@ -1,0 +1,24 @@
+package com.supplyhub.domain.employee.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+
+public record CreateDataEmployeeDto(
+        @NotBlank(message = "Username is required")
+        @Size(min = 3, max = 20, message = "Username must be between 3 and 20 characters")
+        String username,
+
+        @NotBlank(message = "Email is required")
+        @Email(message = "Please provide a valid email address")
+        String email,
+
+        @NotBlank(message = "Password is required")
+        @Size(min = 6, max = 30, message = "Password must be between 6 and 30 characters")
+        String password,
+
+        @NotBlank(message = "CPF is required")
+        @Size(min = 11, max = 11, message = "CPF must be 11 characters")
+        String cpf
+        ) {
+}
