+++ src/main/java/com/supplyhub/domain/employee/dto/DetailsDataEmployeeDto.java
@@ -1,0 +1,17 @@
+package com.supplyhub.domain.employee.dto;
+
+import com.supplyhub.domain.employee.Employee;
+
+public record DetailsDataEmployeeDto(
+        String username,
+        String email,
+        String password
+) {
+    public static DetailsDataEmployeeDto fromEmployee(Employee employee) {
+        return new DetailsDataEmployeeDto(
+                employee.getUsername(),
+                employee.getEmail(),
+                employee.getPassword()
+        );
+    }
+}
