+++ src/main/java/com/supplyhub/domain/employee/dto/EmployeeResponseDto.java
@@ -1,0 +1,13 @@
+package com.supplyhub.domain.employee.dto;
+
+import com.supplyhub.domain.employee.Employee;
+
+public class EmployeeResponseDto {
+    private Long id;
+    private String username;
+
+    public EmployeeResponseDto(Employee employee) {
+        this.id = employee.getId();
+        this.username = employee.getUsername();
+    }
+}
