+++ src/main/java/com/supplyhub/domain/employee/dto/UpdateDataEmployeeData.java
@@ -1,0 +1,12 @@
+package com.supplyhub.domain.employee.dto;
+
+import jakarta.validation.constraints.NotNull;
+
+public record UpdateDataEmployeeData(
+        @NotNull
+        Long id,
+        String username,
+        String email,
+        String password
+) {
+}
