+++ src/main/java/com/supplyhub/domain/user/auth/dto/AuthenticationRequestDto.java
@@ -1,0 +1,7 @@
+package com.supplyhub.domain.user.auth.dto;
+
+public record AuthenticationRequestDto(
+        String email,
+        String password
+) {
+}
