+++ src/main/java/com/supplyhub/domain/user/auth/dto/AuthenticationResponseDto.java
@@ -1,0 +1,4 @@
+package com.supplyhub.domain.user.auth.dto;
+
+public record AuthenticationResponseDto(String token) {
+}
