+++ src/main/java/com/supplyhub/domain/user/dto/UserProfileDto.java
@@ -1,0 +1,12 @@
+package com.supplyhub.domain.user.dto;
+
+import com.supplyhub.domain.user.User;
+
+public record UserProfileDto(String email, String username) {
+    public static UserProfileDto fromUser(User user) {
+        return new UserProfileDto(
+                user.getEmail(),
+                user.getUsername()
+        );
+    }
+}
