--- src/main/java/com/supplyhub/dto/auth/AuthenticationRequestDto.java
@@ -1,7 +1,0 @@
-package com.supplyhub.dto.auth;
-
-public record AuthenticationRequestDto(
-        String email,
-        String password
-) {
-}
