--- src/main/java/com/supplyhub/dto/auth/AuthenticationResponseDto.java
@@ -1,4 +1,0 @@
-package com.supplyhub.dto.auth;
-
-public record AuthenticationResponseDto(String token) {
-}
