--- src/main/java/com/supplyhub/dto/employee/UpdateDataEmployeeData.java
@@ -1,12 +1,0 @@
-package com.supplyhub.dto.employee;
-
-import jakarta.validation.constraints.NotNull;
-
-public record UpdateDataEmployeeData(
-        @NotNull
-        Long id,
-        String username,
-        String email,
-        String password
-) {
-}
