--- src/main/java/com/supplyhub/dto/product/ProductResponseDataDto.java
@@ -1,4 +1,0 @@
-package com.supplyhub.dto.product;
-
-public class ProductResponseDataDto {
-}
