+++ src/main/java/com/supplyhub/infra/security/JwtConfig.java
@@ -3,7 +3,6 @@ package com.supplyhub.infra.security;
 import com.nimbusds.jose.jwk.JWKSet;
 import com.nimbusds.jose.jwk.RSAKey;
 import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
-import com.supplyhub.services.Auth.JwtService;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Value;
