--- src/main/java/com/supplyhub/repositories/EmployeeRepository.java
@@ -1,11 +1,0 @@
-package com.supplyhub.repositories;
-
-import com.supplyhub.entities.Employee;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-    Employee findByCpf(String cpf);
-}
