--- src/main/java/com/supplyhub/repositories/UserRepository.java
@@ -1,16 +1,0 @@
-package com.supplyhub.repositories;
-
-import com.supplyhub.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, UUID> {
-    User findByEmail(String email);
-    Optional<User> findByUsername(String username);
-    boolean existsByUsername(String username);
-    boolean existsByEmail(String email);
-}
