--- src/main/java/com/supplyhub/services/Auth/JpaUserDetailsService.java
@@ -1,30 +1,0 @@
-package com.supplyhub.services.Auth;
-
-import com.supplyhub.repositories.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class JpaUserDetailsService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(final String username)
-            throws UsernameNotFoundException {
-
-        return userRepository.findByUsername(username).map(user ->
-                User.builder()
-                        .username(username)
-                        .password(user.getPassword())
-                        .build()
-        ).orElseThrow(() -> new UsernameNotFoundException(
-                "User with username [%s] not found".formatted(username)));
-    }
-
-}
