+++ src/main/resources/application.properties
@@ -8,12 +8,15 @@ spring.jpa.properties.hibernate.format_sql=false
 
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
-spring.mail.username=
+spring.mail.username=Jamilla Lobo
 spring.mail.password=
 spring.mail.properties.mail.transport.protocol=smtp
 spring.mail.properties.mail.smtp.starttls.enable=true
 spring.mail.properties.mail.smtp.starttls.required=true
 spring.mail.properties.mail.smtp.auth=true
-spring.mail.properties.mail.smtp.from=exemplo@gmail.com
+spring.mail.properties.mail.smtp.from=jamillalobo@gmail.com
 spring.mail.default-encoding=UTF-8
 spring.mail.test-connection=false
+
+logging.level.org.springframework.web=DEBUG
+logging.level.org.springdoc=DEBUG
