+++ src/main/resources/db/migration/V3__Alter_products_table.sql.sql
